#423: 투두 조회 presentation 로직 추가
- EditToDoPresenter에 투두 조회 Presentation 로직을 추가하였습니다.

+++ ToDo-Garden/ToDo-Garden/EditToDoScene/Sources/EditToDoSceneEntity/EditToDoModels.swift
@@ -292,9 +292,9 @@ extension EditToDo {
 
   public struct ToDoAlarm {
     public var isAlarmOn: Bool
-    public var alarmTime: Int?
+    public var alarmTime: Double?
 
-    public init(isAlarmOn: Bool, alarmTime: Int?) {
+    public init(isAlarmOn: Bool, alarmTime: Double?) {
       self.isAlarmOn = isAlarmOn
       self.alarmTime = alarmTime
     }
